feat: add new elements
